feat: lista dois revisada

+++ Listas de Exercícios/LP_Lista_02/ex03/src/Main.java
@@ -3,12 +3,12 @@ import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
-        int number, result;
+        double number, result;
 
         System.out.println("Digite um número real: ");
 
         Scanner in = new Scanner(System.in); //Configura o scanner
-        number = in.nextInt();
+        number = in.nextDouble();
 
         result = number / 5;
         System.out.println("A quinta parte de " + number + " é " + result);
